+++ x/zoneconcierge/keeper/keeper.go
@@ -5,10 +5,11 @@ import (
 	corestoretypes "cosmossdk.io/core/store"
 	"cosmossdk.io/log"
 	storetypes "cosmossdk.io/store/types"
-	"github.com/babylonlabs-io/babylon/v3/x/zoneconcierge/types"
 	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	ibcexported "github.com/cosmos/ibc-go/v10/modules/core/exported"
+
+	"github.com/babylonlabs-io/babylon/v3/x/zoneconcierge/types"
 )
 
 type (
@@ -34,6 +35,9 @@ type (
 		// Typically, this should be the x/gov module account.
 		authority string
 
+		// Transient store key for tracking BTC header and consumer event broadcasting triggers
+		transientKey *storetypes.TransientStoreKey
+
 		// Collections for KV store management
 		Schema                collections.Schema
 		ParamsCollection      collections.Item[types.Params]
@@ -47,6 +51,7 @@ type (
 func NewKeeper(
 	cdc codec.BinaryCodec,
 	storeService corestoretypes.KVStoreService,
+	transientKey *storetypes.TransientStoreKey,
 	ics4Wrapper types.ICS4Wrapper,
 	clientKeeper types.ClientKeeper,
 	connectionKeeper types.ConnectionKeeper,
@@ -81,6 +86,7 @@ func NewKeeper(
 		bsKeeper:            bsKeeper,
 		btcStkKeeper:        btcStkKeeper,
 		authority:           authority,
+		transientKey:        transientKey,
 
 		ParamsCollection: collections.NewItem[types.Params](
 			sb,